// src/common/interfaces/subscriber-pattern.interface.ts
export enum SubscriberPattern {
  CREATE_NEW_USER = 'create_new_user',
  CREATE_NEW_ADMIN_USER = 'create_new_admin_user',
  LOGIN = 'login',
  ACCOUNT_ACTIVATION = 'account_activation',
  FORGOT_PASSWORD = 'forgot_password',
  RESET_PASSWORD = 'reset_password',
  USER_PROFILE = 'user_profile',
  UPDATE_USER = 'update_user',
}